Extract host IO logic from GC Heap to host module
