+++ .jules/atlas.md
@@ -12,3 +12,6 @@
 **[Split ClassFile types]
 **Tangle:** `duke-classfile/src/types.rs` was a Blob anti-pattern holding constant pool, attributes, and class structures.
 **Blueprint:** Split into `constant_pool.rs`, `attributes.rs`, and `class.rs` to match domain responsibilities.
+**[Extract Host IO from GC Heap]
+**Tangle:** The `duke-gc/src/lib.rs` file contained native OS file and process handling mixed directly with the generational garbage collector logic, forming a large Blob.
+**Blueprint:** Extracted the `HostFileHandle`, `HostProcessHandle`, and all native OS interaction methods from `Heap` into a dedicated `host` module to cleanly separate I/O from memory management.
